Move all parsers in one file `parsers.py`
