+++ apps/parse/management/commands/parse.py
@@ -1,6 +1,6 @@
 from django.core.management.base import BaseCommand, CommandParser
 
-from apps.parse.readmanga import parser
+from apps.parse import parsers
 
 SETTINGS_PATH = "apps.parse.readmanga.readmanga.settings"
 
@@ -21,7 +21,7 @@ class Command(BaseCommand):
     def handle(self, *args, **options):
         parser_name = options.get("parser")
         try:
-            getattr(parser, f"{parser_name}_parser")()
+            getattr(parsers, f"{parser_name}_parser")()
         except AttributeError:
             self.stdout.write(f"Can't find parser [{parser_name}]", self.style.ERROR)
         except Exception:
